+++ src/app/components/component.page.js
@@ -1,4 +1,6 @@
+import React from 'react';
 import PropTypes from 'prop-types';
+
 import SocketComponent from './component.socket';
 
 /**
@@ -29,6 +31,22 @@ export default class Page extends SocketComponent {
 		if (this.getContext().DEV) {
 			console.log(`Page '${this.constructor.name}' will unmount`);
 		}
+	}
+
+	render() {
+		if (!this.getContext().DEV) {
+			return null;
+		}
+		return (
+			<div
+				style={{
+					padding: 10,
+					backgroundColor: '#F0F0F0',
+				}}
+			>
+				<h1 style={{ textAlign: 'center' }}>{this.constructor.name}</h1>
+			</div>
+		);
 	}
 
 	onPageRender() { }